Split with GrowthCampaign HOC
